+++ crates/jsonschema-py/src/lib.rs
@@ -412,6 +412,7 @@ fn make_options(
     ignore_unknown_formats: Option<bool>,
     retriever: Option<&Bound<'_, PyAny>>,
     registry: Option<&registry::Registry>,
+    base_uri: Option<String>,
 ) -> PyResult<jsonschema::ValidationOptions> {
     let mut options = jsonschema::options();
     if let Some(raw_draft_version) = draft {
@@ -460,6 +461,9 @@ fn make_options(
     }
     if let Some(registry) = registry {
         options = options.with_registry(registry.inner.clone());
+    }
+    if let Some(base_uri) = base_uri {
+        options = options.with_base_uri(base_uri);
     }
     Ok(options)
 }
@@ -570,7 +574,7 @@ impl Write for StringWriter<'_> {
     }
 }
 
-/// is_valid(schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None)
+/// is_valid(schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None, base_uri=None)
 ///
 /// A shortcut for validating the input instance against the schema.
 ///
@@ -580,7 +584,7 @@ impl Write for StringWriter<'_> {
 /// If your workflow implies validating against the same schema, consider using `validator_for(...).is_valid`
 /// instead.
 #[pyfunction]
-#[pyo3(signature = (schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None))]
+#[pyo3(signature = (schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None, base_uri=None))]
 fn is_valid(
     py: Python<'_>,
     schema: &Bound<'_, PyAny>,
@@ -592,6 +596,7 @@ fn is_valid(
     retriever: Option<&Bound<'_, PyAny>>,
     registry: Option<&registry::Registry>,
     mask: Option<String>,
+    base_uri: Option<String>,
 ) -> PyResult<bool> {
     let options = make_options(
         draft,
@@ -600,6 +605,7 @@ fn is_valid(
         ignore_unknown_formats,
         retriever,
         registry,
+        base_uri,
     )?;
     let schema = ser::to_value(schema)?;
     match options.build(&schema) {
@@ -612,7 +618,7 @@ fn is_valid(
     }
 }
 
-/// validate(schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None)
+/// validate(schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None, base_uri=None)
 ///
 /// Validate the input instance and raise `ValidationError` in the error case
 ///
@@ -624,7 +630,7 @@ fn is_valid(
 /// If your workflow implies validating against the same schema, consider using `validator_for(...).validate`
 /// instead.
 #[pyfunction]
-#[pyo3(signature = (schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None))]
+#[pyo3(signature = (schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None, base_uri=None))]
 fn validate(
     py: Python<'_>,
     schema: &Bound<'_, PyAny>,
@@ -636,6 +642,7 @@ fn validate(
     retriever: Option<&Bound<'_, PyAny>>,
     registry: Option<&registry::Registry>,
     mask: Option<String>,
+    base_uri: Option<String>,
 ) -> PyResult<()> {
     let options = make_options(
         draft,
@@ -644,6 +651,7 @@ fn validate(
         ignore_unknown_formats,
         retriever,
         registry,
+        base_uri,
     )?;
     let schema = ser::to_value(schema)?;
     match options.build(&schema) {
@@ -652,7 +660,7 @@ fn validate(
     }
 }
 
-/// iter_errors(schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None)
+/// iter_errors(schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None, base_uri=None)
 ///
 /// Iterate the validation errors of the input instance
 ///
@@ -663,7 +671,7 @@ fn validate(
 /// If your workflow implies validating against the same schema, consider using `validator_for().iter_errors`
 /// instead.
 #[pyfunction]
-#[pyo3(signature = (schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None))]
+#[pyo3(signature = (schema, instance, draft=None, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None, base_uri=None))]
 fn iter_errors(
     py: Python<'_>,
     schema: &Bound<'_, PyAny>,
@@ -675,6 +683,7 @@ fn iter_errors(
     retriever: Option<&Bound<'_, PyAny>>,
     registry: Option<&registry::Registry>,
     mask: Option<String>,
+    base_uri: Option<String>,
 ) -> PyResult<ValidationErrorIter> {
     let options = make_options(
         draft,
@@ -683,6 +692,7 @@ fn iter_errors(
         ignore_unknown_formats,
         retriever,
         registry,
+        base_uri,
     )?;
     let schema = ser::to_value(schema)?;
     match options.build(&schema) {
@@ -708,7 +718,7 @@ struct Validator {
     mask: Option<String>,
 }
 
-/// validator_for(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None)
+/// validator_for(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None, base_uri=None)
 ///
 /// Create a validator for the input schema with automatic draft detection and default options.
 ///
@@ -717,7 +727,7 @@ struct Validator {
 ///     False
 ///
 #[pyfunction]
-#[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None))]
+#[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None, base_uri=None))]
 fn validator_for(
     py: Python<'_>,
     schema: &Bound<'_, PyAny>,
@@ -727,6 +737,7 @@ fn validator_for(
     retriever: Option<&Bound<'_, PyAny>>,
     registry: Option<&registry::Registry>,
     mask: Option<String>,
+    base_uri: Option<String>,
 ) -> PyResult<Validator> {
     validator_for_impl(
         py,
@@ -738,6 +749,7 @@ fn validator_for(
         retriever,
         registry,
         mask,
+        base_uri,
     )
 }
 
@@ -751,6 +763,7 @@ fn validator_for_impl(
     retriever: Option<&Bound<'_, PyAny>>,
     registry: Option<&registry::Registry>,
     mask: Option<String>,
+    base_uri: Option<String>,
 ) -> PyResult<Validator> {
     let obj_ptr = schema.as_ptr();
     let object_type = unsafe { pyo3::ffi::Py_TYPE(obj_ptr) };
@@ -770,6 +783,7 @@ fn validator_for_impl(
         ignore_unknown_formats,
         retriever,
         registry,
+        base_uri,
     )?;
     match options.build(&schema) {
         Ok(validator) => Ok(Validator { validator, mask }),
@@ -780,7 +794,7 @@ fn validator_for_impl(
 #[pymethods]
 impl Validator {
     #[new]
-    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry = None, mask=None))]
+    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry = None, mask=None, base_uri=None))]
     fn new(
         py: Python<'_>,
         schema: &Bound<'_, PyAny>,
@@ -790,6 +804,7 @@ impl Validator {
         retriever: Option<&Bound<'_, PyAny>>,
         registry: Option<&registry::Registry>,
         mask: Option<String>,
+        base_uri: Option<String>,
     ) -> PyResult<Self> {
         validator_for(
             py,
@@ -800,6 +815,7 @@ impl Validator {
             retriever,
             registry,
             mask,
+            base_uri,
         )
     }
     /// is_valid(instance)
@@ -859,7 +875,7 @@ impl Validator {
     }
 }
 
-/// Draft4Validator(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None)
+/// Draft4Validator(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None, base_uri=None)
 ///
 /// A JSON Schema Draft 4 validator.
 ///
@@ -873,7 +889,7 @@ struct Draft4Validator {}
 #[pymethods]
 impl Draft4Validator {
     #[new]
-    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None))]
+    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None, base_uri=None))]
     fn new(
         py: Python<'_>,
         schema: &Bound<'_, PyAny>,
@@ -883,6 +899,7 @@ impl Draft4Validator {
         retriever: Option<&Bound<'_, PyAny>>,
         registry: Option<&registry::Registry>,
         mask: Option<String>,
+        base_uri: Option<String>,
     ) -> PyResult<(Self, Validator)> {
         Ok((
             Draft4Validator {},
@@ -896,12 +913,13 @@ impl Draft4Validator {
                 retriever,
                 registry,
                 mask,
+                base_uri,
             )?,
         ))
     }
 }
 
-/// Draft6Validator(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None)
+/// Draft6Validator(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None, base_uri=None)
 ///
 /// A JSON Schema Draft 6 validator.
 ///
@@ -915,7 +933,7 @@ struct Draft6Validator {}
 #[pymethods]
 impl Draft6Validator {
     #[new]
-    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None))]
+    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None, base_uri=None))]
     fn new(
         py: Python<'_>,
         schema: &Bound<'_, PyAny>,
@@ -925,6 +943,7 @@ impl Draft6Validator {
         retriever: Option<&Bound<'_, PyAny>>,
         registry: Option<&registry::Registry>,
         mask: Option<String>,
+        base_uri: Option<String>,
     ) -> PyResult<(Self, Validator)> {
         Ok((
             Draft6Validator {},
@@ -938,12 +957,13 @@ impl Draft6Validator {
                 retriever,
                 registry,
                 mask,
+                base_uri,
             )?,
         ))
     }
 }
 
-/// Draft7Validator(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None)
+/// Draft7Validator(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None, base_uri=None)
 ///
 /// A JSON Schema Draft 7 validator.
 ///
@@ -957,7 +977,7 @@ struct Draft7Validator {}
 #[pymethods]
 impl Draft7Validator {
     #[new]
-    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None))]
+    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None, base_uri=None))]
     fn new(
         py: Python<'_>,
         schema: &Bound<'_, PyAny>,
@@ -967,6 +987,7 @@ impl Draft7Validator {
         retriever: Option<&Bound<'_, PyAny>>,
         registry: Option<&registry::Registry>,
         mask: Option<String>,
+        base_uri: Option<String>,
     ) -> PyResult<(Self, Validator)> {
         Ok((
             Draft7Validator {},
@@ -980,12 +1001,13 @@ impl Draft7Validator {
                 retriever,
                 registry,
                 mask,
+                base_uri,
             )?,
         ))
     }
 }
 
-/// Draft201909Validator(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None)
+/// Draft201909Validator(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None, base_uri=None)
 ///
 /// A JSON Schema Draft 2019-09 validator.
 ///
@@ -999,7 +1021,7 @@ struct Draft201909Validator {}
 #[pymethods]
 impl Draft201909Validator {
     #[new]
-    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None))]
+    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None, base_uri=None))]
     fn new(
         py: Python<'_>,
         schema: &Bound<'_, PyAny>,
@@ -1009,6 +1031,7 @@ impl Draft201909Validator {
         retriever: Option<&Bound<'_, PyAny>>,
         registry: Option<&registry::Registry>,
         mask: Option<String>,
+        base_uri: Option<String>,
     ) -> PyResult<(Self, Validator)> {
         Ok((
             Draft201909Validator {},
@@ -1022,12 +1045,13 @@ impl Draft201909Validator {
                 retriever,
                 registry,
                 mask,
+                base_uri,
             )?,
         ))
     }
 }
 
-/// Draft202012Validator(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None)
+/// Draft202012Validator(schema, formats=None, validate_formats=None, ignore_unknown_formats=True, retriever=None, registry=None, mask=None, base_uri=None)
 ///
 /// A JSON Schema Draft 2020-12 validator.
 ///
@@ -1041,7 +1065,7 @@ struct Draft202012Validator {}
 #[pymethods]
 impl Draft202012Validator {
     #[new]
-    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None))]
+    #[pyo3(signature = (schema, formats=None, validate_formats=None, ignore_unknown_formats=true, retriever=None, registry=None, mask=None, base_uri=None))]
     fn new(
         py: Python<'_>,
         schema: &Bound<'_, PyAny>,
@@ -1051,6 +1075,7 @@ impl Draft202012Validator {
         retriever: Option<&Bound<'_, PyAny>>,
         registry: Option<&registry::Registry>,
         mask: Option<String>,
+        base_uri: Option<String>,
     ) -> PyResult<(Self, Validator)> {
         Ok((
             Draft202012Validator {},
@@ -1064,6 +1089,7 @@ impl Draft202012Validator {
                 retriever,
                 registry,
                 mask,
+                base_uri,
             )?,
         ))
     }